Fix links on the Homepage
Links were not being redirected to the correct page due to a bad href
value attribution. Instead of working as an relative reference and
looking for the page in the recipes directory, the links were being read
as an absolute reference due to the lack of a "./" before the directory
name.